added files/memory cleaning, fixed bugs

+++ srcs/http/Message.hpp
@@ -77,7 +77,7 @@ namespace HTTP
 			std::string	body;
 			friend class Client;
 		private:
-			int		content_length;
+			size_t	content_length;
 			size_t	header_bytes;
 			std::map<std::string, std::string>	headers; //Max size of the headers section: 8k
 			std::vector<std::string>			method; //Max size: 8k
